+++ lib/api/src/grpc/proto/collections.proto
@@ -310,7 +310,7 @@ enum ShardingMethod {
   Custom = 1; // Shard by user-defined key
 }
 
-message StrictModeConfigDiff {
+message StrictModeConfig {
   optional bool enabled = 1;
   optional uint32 max_query_limit = 2;
   optional uint32 max_timeout = 3;
@@ -339,7 +339,7 @@ message CreateCollection {
   optional QuantizationConfig quantization_config = 14; // Quantization configuration of vector
   optional ShardingMethod sharding_method = 15; // Sharding method
   optional SparseVectorConfig sparse_vectors_config = 16; // Configuration for sparse vectors
-  optional StrictModeConfigDiff strict_mode_config = 17; // Configuration for strict mode
+  optional StrictModeConfig strict_mode_config = 17; // Configuration for strict mode
 }
 
 message UpdateCollection {
@@ -389,7 +389,7 @@ message CollectionConfig {
   OptimizersConfigDiff optimizer_config = 3; // Configuration of the optimizers
   WalConfigDiff wal_config = 4; // Configuration of the Write-Ahead-Log
   optional QuantizationConfig quantization_config = 5; // Configuration of the vector quantization
-  optional StrictModeConfigDiff strict_mode_config = 6; // Configuration of strict mode.
+  optional StrictModeConfig strict_mode_config = 6; // Configuration of strict mode.
 }
 
 enum TokenizerType {
