+++ lib/segment/Cargo.toml
@@ -99,6 +99,7 @@ macro_rules_attribute = "0.2.0"
 generic-tests = { workspace = true }
 nom = "7.1.3"
 half = { workspace = true }
+merge = { workspace = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 cgroups-rs = "0.3"
